refactor(admin): extract SocialLinksIcons from RosterTab, pin behaviour with tests
Pure move, no logic change. The Links column had no test coverage at all
(the one e2e test that covered it is test.skip'd), so it is characterized
here before the registry refactor rewrites its internals.

Also the first test to actually invoke the SocialIcons components, which
RosterTab.test.jsx has been loading into the coverage denominator without
ever rendering.
